Add files via upload
This notebook classifies movie reviews as positive or negative using the text of the review. 
We'll use the IMDB dataset that contains the text of 50,000 movie reviews from the Internet Movie Database. 
These are split into 25,000 reviews for training and 25,000 reviews for testing. 
The training and testing sets are balanced, meaning they contain an equal number of positive and negative reviews.